icetime: never silently truncate asc file lines
icetime was reading the asc configuration file using a 128-byte line
buffer -- which is usually fine, but can cause it to truncate the names
of nets given in .sym lines if those names are very, very long.  The way
fgets was being used meant this went undetected.

Long net names like this can arise in deeply hierarchical designs,
particularly if there's a code generator involved.

This change:
1. Increases the buffer size to 64kiB.
2. Adds a truncation check that causes icetime to fail.

A more robust solution would manage the line buffer on the heap, since
the symbol gets copied into a std::string anyway, but this is a
workaround for now.