Fix case where make_prim allocates all global buffer pins
This is a low probability bug more likely to show up in low pin
count devices with few GBINs. In rare cases make_prim would
constrain all of the global buffer capable pins but not the clock
input. icecube would then fail to place the clock input. This is
fixed by always constraining the clock if all GBIN pins are used.