Replace sed -i with perl for portability
In place editing without a backup is not very portable.
BSD sed requires an extension. To avoid a backup, a zero-length extension must be specified (`sed -i ''`).
GNU sed does not require an extension. To avoid a backup, no extension is specified (`sed -i`).

`sed -i''` and `sed -i` work with GNU sed but not BSD sed.
`sed -i ''` works with BSD sed but not GNU sed.