Speed up Pico programmer by switching to raw USB protocol
This update replaces the HID-based USB protocol with a vendor
control transfer variant. The primary goal is speed- HID transfers
are rate-limited to 1000/s, which gives a maximum transfer rate of
64kb/s. Unfortunately a majority of this bandwidth is wasted, and
typical ice40 flash routines could take up to 17 seconds to
complete. With the switch to a new protocol, the time is reduced to
under 4 seconds.